Preparing to organize project into something 'Manager' oriented, and have begun work on the TileMap class and RenderManager.

+++ src/dgk/prototype/game/Block.java
@@ -12,7 +12,7 @@ public class Block extends BuildingComponent {
 
     private boolean isPassable;
 
-    public Block(int textureId, byte renderLayer, double x, double y) {
+    public Block(int textureId, byte renderLayer, int x, int y) {
         super(textureId, x, y, World.GRID_SIZE, World.GRID_SIZE);
 
         this.isPassable = true;
